+++ R/map_layer_arc.R
@@ -129,7 +129,7 @@ mapdeckArcDependency <- function() {
 #' url <- 'https://raw.githubusercontent.com/plotly/datasets/master/2011_february_aa_flight_paths.csv'
 #' flights <- read.csv(url)
 #' flights$id <- seq_len(nrow(flights))
-#' flights$stroke <- sample(1:3, size = nrow(flights), replace = T)
+#' flights$stroke <- sample(1:3, size = nrow(flights), replace = TRUE)
 #' flights$info <- paste0("<b>",flights$airport1, " - ", flights$airport2, "</b>")
 #'
 #' mapdeck( style = mapdeck_style("dark"), pitch = 45 ) %>%
@@ -143,7 +143,7 @@ mapdeckArcDependency <- function() {
 #'   , stroke_width = "stroke"
 #'   , tooltip = "info"
 #'   , auto_highlight = TRUE
-#'   , legend = T
+#'   , legend = TRUE
 #'   , legend_options = list(
 #'     stroke_from = list( title = "Origin airport" ),
 #'     css = "max-height: 100px;")