+++ R/map_layer_grid.R
@@ -65,7 +65,7 @@ mapdeckGridDependency <- function() {
 #' )
 #'
 #' ## using colour and elevation functions, and legends
-#' df$val <- sample(1:10, size = nrow(df), replace = T)
+#' df$val <- sample(1:10, size = nrow(df), replace = TRUE)
 #'
 #' mapdeck( style = mapdeck_style("dark"), pitch = 45) %>%
 #' add_grid(
@@ -74,7 +74,7 @@ mapdeckGridDependency <- function() {
 #' 	, lon = "lng"
 #' 	, layer_id = "hex_layer"
 #' 	, elevation_scale = 100
-#' 	, legend = T
+#' 	, legend = TRUE
 #' 	, colour_function = "max"
 #' 	, colour = "val"
 #' )
@@ -86,7 +86,7 @@ mapdeckGridDependency <- function() {
 #' 	, lon = "lng"
 #' 	, layer_id = "hex_layer"
 #' 	, elevation_scale = 10
-#' 	, legend = T
+#' 	, legend = TRUE
 #' 	, elevation_function = "mean"
 #' 	, elevation = "val"
 #' )