+++ R/mapdeck_dependencies.R
@@ -155,13 +155,12 @@ mapdeck_dep_location <- function() {
 	)
 }
 
-
 # https://unpkg.com/deck.gl@latest/dist.min.js
 deckgl_min_js <- function() {
 	list(
 		createHtmlDependency(
 			name = "deckgl",
-			version = "8.1.1",
+			version = "8.2.9",
 			src = system.file("htmlwidgets/lib/", package = "mapdeck"),
 			script = c("deckgl.min.js"),
 			all_files = FALSE
@@ -169,13 +168,13 @@ deckgl_min_js <- function() {
 	)
 }
 
-## https://api.mapbox.com/mapbox-gl-js/v1.9.1/mapbox-gl.js
-## https://api.mapbox.com/mapbox-gl-js/v1.9.1/mapbox-gl.css
+## https://api.mapbox.com/mapbox-gl-js/v1.12.0/mapbox-gl.js
+## https://api.mapbox.com/mapbox-gl-js/v1.12.0/mapbox-gl.css
 mapboxgl <- function() {
 	list(
 		createHtmlDependency(
 			name = "mapboxgl",
-			version = "1.9.1",
+			version = "1.12.0",
 			src = system.file("htmlwidgets/lib/", package = "mapdeck"),
 			script = c("mapbox-gl.js"),
 			stylesheet = c("mapbox-gl.css"),