+++ R/mapdeck_map_utilities.R
@@ -122,7 +122,8 @@ doResolveFormula.data.frame = function(data, f) {
 # @param layer_id
 layerId <- function(
 	layer_id,
-	layer = c("animated_arc", "animated_line", "arc", "bitmap", "column", "geojson", "greatcircle","grid","heatmap","hexagon",
+	layer = c("animated_arc", "animated_line", "arc", "bitmap", "cesium", "column",
+						"geojson", "greatcircle","grid","heatmap","hexagon", "i3s",
 						"line", "mesh", "path","pointcloud", "polygon","scatterplot", "screengrid", "terrain",
 						"text", "title","trips")
 	) {
@@ -136,4 +137,10 @@ layerId <- function(
 	} else {
 		return(layer_id)
 	}
+}
+
+# returns 0-based index of all the list columns in a data.frame (sf object)
+list_columns <- function(x, geometry_col) {
+	geom_col <- which(names(x) == geometry_col)
+	return( setdiff( which( vapply(x, is.list, T) ), geom_col ) - 1 )
 }