+++ src/geojson.cpp
@@ -24,7 +24,7 @@ Rcpp::List rcpp_geojson_geojson(
 		int digits
 	) {
 
-	int data_rows = data.nrows();
+	R_xlen_t data_rows = data.nrows();
 
 	Rcpp::List lst_defaults = geojson_defaults( data_rows );  // initialise with defaults